Move configTypeToFileSuffix to utils
